fixed problem on iOS7.1 64bit

+++ Spika/Views/MessageTableViewCells/MessageTypeBasicCell.m
@@ -144,7 +144,7 @@
 
 #pragma mark - Override
 
-+ (CGFloat) cellHeightForMessage:(ModelMessage *)message{
++ (float) cellHeightForMessage:(ModelMessage *)message{
     
     NSAssert(NO, @"You have to override this method!");
     return 0;
